Struct priority_queue::priority_queue::PriorityQueue

pub struct PriorityQueue<I, P, H = RandomState>where
    I: Hash + Eq,
    P: Ord,{ /* private fields */ }

A priority queue with efficient change function to change the priority of an element.

The priority is of type P, that must implement std::cmp::Ord.

The item is of type I, that must implement Hash and Eq.

Implemented as a heap of indexes, stores the items inside an IndexMap to be able to retrieve them quickly.

Example

use priority_queue::PriorityQueue;

let mut pq = PriorityQueue::new();

assert!(pq.is_empty());
pq.push("Apples", 5);
pq.push("Bananas", 8);
pq.push("Strawberries", 23);

assert_eq!(pq.peek(), Some((&"Strawberries", &23)));

pq.change_priority("Bananas", 25);
assert_eq!(pq.peek(), Some((&"Bananas", &25)));

for (item, _) in pq.into_sorted_iter() {
    println!("{}", item);
}

Implementations

impl<I, P> PriorityQueue<I, P>where
P: Ord,
I: Hash + Eq,

pub fn new() -> Self

Creates an empty PriorityQueue

pub fn with_capacity(capacity: usize) -> Self

Creates an empty PriorityQueue with the specified capacity.

impl<I, P, H> PriorityQueue<I, P, H>where
P: Ord,
I: Hash + Eq,
H: BuildHasher + Default,

pub fn with_default_hasher() -> Self

Creates an empty PriorityQueue with the default hasher

pub fn with_capacity_and_default_hasher(capacity: usize) -> Self

Creates an empty PriorityQueue with the specified capacity and default hasher

impl<I, P, H> PriorityQueue<I, P, H>where
P: Ord,
I: Hash + Eq,
H: BuildHasher,

pub fn with_hasher(hash_builder: H) -> Self

Creates an empty PriorityQueue with the specified hasher

pub fn with_capacity_and_hasher(capacity: usize, hash_builder: H) -> Self

Creates an empty PriorityQueue with the specified capacity and hasher

The internal collections will be able to hold at least capacity elements without reallocating. If capacity is 0, there will be no allocation.

pub fn iter(&self) -> Iter<'_, I, P>

Returns an iterator in arbitrary order over the (item, priority) elements in the queue

impl<I, P, H> PriorityQueue<I, P, H>where
P: Ord,
I: Hash + Eq,

pub fn iter_mut(&mut self) -> IterMut<'_, I, P, H>

Returns an iterator in arbitrary order over the (item, priority) elements in the queue.

The item and the priority are mutable references, but it’s a logic error to modify the item in a way that change the result of Hash or Eq.

It’s not an error, instead, to modify the priorities, because the heap will be rebuilt once the IterMut goes out of scope. It would be rebuilt even if no priority value would have been modified, but the procedure will not move anything, but just compare the priorities.

pub fn peek(&self) -> Option<(&I, &P)>

Returns the couple (item, priority) with the greatest priority in the queue, or None if it is empty.

Computes in O(1) time

pub fn peek_mut(&mut self) -> Option<(&mut I, &P)>

Returns the couple (item, priority) with the greatest priority in the queue, or None if it is empty.

The item is a mutable reference, but it’s a logic error to modify it in a way that change the result of Hash or Eq.

The priority cannot be modified with a call to this function. To modify the priority use push, change_priority or change_priority_by.

Computes in O(1) time

pub fn capacity(&self) -> usize

Returns the number of elements the internal map can hold without reallocating.

This number is a lower bound; the map might be able to hold more, but is guaranteed to be able to hold at least this many.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the internal data structures that support this operation as much as possible.

pub fn pop(&mut self) -> Option<(I, P)>

Removes the item with the greatest priority from the priority queue and returns the pair (item, priority), or None if the queue is empty.

pub fn into_sorted_vec(self) -> Vec<I>

Implements a HeapSort.

Returns a Vec<I> sorted from the item associated to the highest priority to the lowest.

pub fn len(&self) -> usize

Returns the number of elements in the priority queue.

pub fn is_empty(&self) -> bool

Returns true if the priority queue contains no elements.

pub fn into_sorted_iter(self) -> IntoSortedIter<I, P, H>

Generates a new iterator from self that will extract the elements from the one with the highest priority to the lowest one.

impl<I, P, H> PriorityQueue<I, P, H>where
P: Ord,
I: Hash + Eq,
H: BuildHasher,

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the given PriorityQueue. The collection may reserve more space to avoid frequent reallocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

Panics

Panics if the new capacity overflows usize.

pub fn push(&mut self, item: I, priority: P) -> Option<P>

Insert the item-priority pair into the queue.

If an element equal to item was already into the queue, it is updated and the old value of its priority returned in Some; otherwise, returns None.

Computes in O(log(N)) time.

pub fn push_increase(&mut self, item: I, priority: P) -> Option<P>

Increase the priority of an existing item in the queue, or insert it if not present.

If an element equal to item is already in the queue with a lower priority, its priority is increased to the new one without replacing the element and the old priority is returned. Otherwise, the new element is inserted into the queue.

Returns Some if an element equal to item is already in the queue. If its priority is higher then priority, the latter is returned back, otherwise, the old priority is contained in the Option. If the item is not in the queue, None is returned.

Computes in O(log(N)) time.

pub fn push_decrease(&mut self, item: I, priority: P) -> Option<P>

Decrease the priority of an existing item in the queue, or insert it if not present.

If an element equal to item is already in the queue with a higher priority, its priority is decreased to the new one without replacing the element and the old priority is returned. Otherwise, the new element is inserted into the queue.

Returns Some if an element equal to item is already in the queue. If its priority is lower then priority, the latter is returned back, otherwise, the old priority is contained in the Option. If the item is not in the queue, None is returned.

Computes in O(log(N)) time.

pub fn change_priority<Q>(&mut self, item: &Q, new_priority: P) -> Option<P>where
I: Borrow<Q>,
Q: Eq + Hash + ?Sized,

Change the priority of an Item returning the old value of priority, or None if the item wasn’t in the queue.

The argument item is only used for lookup, and is not used to overwrite the item’s data in the priority queue.

The item is found in O(1) thanks to the hash table. The operation is performed in O(log(N)) time.

pub fn change_priority_by<Q, F>(&mut self, item: &Q, priority_setter: F) -> boolwhere
I: Borrow<Q>,
Q: Eq + Hash + ?Sized,
F: FnOnce(&mut P),

Change the priority of an Item using the provided function. Return a boolean value where true means the item was in the queue and update was successful

The argument item is only used for lookup, and is not used to overwrite the item’s data in the priority queue.

The item is found in O(1) thanks to the hash table. The operation is performed in O(log(N)) time (worst case).

pub fn get_priority<Q>(&self, item: &Q) -> Option<&P>where
I: Borrow<Q>,
Q: Eq + Hash + ?Sized,

Get the priority of an item, or None, if the item is not in the queue

pub fn get<Q>(&self, item: &Q) -> Option<(&I, &P)>where
I: Borrow<Q>,
Q: Eq + Hash + ?Sized,

Get the couple (item, priority) of an arbitrary element, as reference or None if the item is not in the queue.

pub fn get_mut<Q>(&mut self, item: &Q) -> Option<(&mut I, &P)>where
I: Borrow<Q>,
Q: Eq + Hash + ?Sized,

Get the couple (item, priority) of an arbitrary element, or None if the item was not in the queue.

The item is a mutable reference, but it’s a logic error to modify it in a way that change the result of Hash or Eq.

The priority cannot be modified with a call to this function. To modify the priority use push, change_priority or change_priority_by.

pub fn remove<Q>(&mut self, item: &Q) -> Option<(I, P)>where
I: Borrow<Q>,
Q: Eq + Hash + ?Sized,

Remove an arbitrary element from the priority queue. Returns the (item, priority) couple or None if the item is not found in the queue.

The operation is performed in O(log(N)) time (worst case).

pub fn into_vec(self) -> Vec<I>

Returns the items not ordered

pub fn clear(&mut self)

Drops all items from the priority queue

pub fn append(&mut self, other: &mut Self)

Move all items of the other queue to self ignoring the items Eq to elements already in self At the end, other will be empty.

Note that at the end, the priority of the duplicated elements inside self may be the one of the elements in other, if other is longer than self